feat: authorization & delete functionality added

+++ App/controllers/ListingController.php
@@ -3,8 +3,10 @@
 namespace App\Controllers;
 
 use ErrorController;
+use Framework\Authorization;
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
 
 class ListingController
 {
@@ -79,7 +81,7 @@ class ListingController
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
         // This is hardcoded for now
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         // Sanitize the data
         $newListingData = array_map('sanitize', $newListingData);
@@ -147,6 +149,13 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('Listing not found!');
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = "You are not authorized to delete this listing";
+
+            return redirect("/listing/{$id}");
         }
 
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
